+++ src/common/constants/options.js
@@ -1,2 +1,5 @@
 export const CUSTOM_DIRECTORY = "custom-directory";
 export const IS_CUSTOM_DIRECTORY = "is-custom-directory";
+
+export const HISTORY_MAX_COUNT = "history-max-count";
+export const HISTORY_EXPIRATION_TIME = "history-expiration-time";